hikey/hikey960: Force navkeys on w/ GSI via qemu.hw.mainkeys=0
It seems a bit odd, but when booting with a GSI build, there
are no navigation keys. Looking at the code, it seems the
qemu.hw.mainkeys property can be used to force it, and other
devices like marlin use this as well.

Many thanks to Vishal Bhoj for helping find/validate this.

Change-Id: I15c625b51a9c9f7d71825e7f2685d4362e57ca82
Signed-off-by: John Stultz <john.stultz@linaro.org>