hikey: Set default kernel to 4.14
We had held back on moving forward to 4.14 as it ran into the
horizontal offset issue more frequently than 4.9 kernels (though
it sometimes would happen w/ 4.9 and 4.4 kernels too).

See: https://bugs.96boards.org/show_bug.cgi?id=671

But after a year of the issue being open, we have not had any
progress from HiSi on it.

The issue there is odd, and doesn't always reproduce, and at this
point I feel that holding back newer kernel support is a bigger
negative then the occasional horizontal offset glitch (which can
be worked around by unplugging and replugging the hdmi cable).

So move forward to 4.14 by default for HiKey, with the expectation
that we will move on to 4.19 ASAP once a few missing features are
fixed there.

Change-Id: I3cd68e65002db48a29c7740264e2c2b28bdc12a3
Signed-off-by: John Stultz <john.stultz@linaro.org>