gralloc960: Clean up previous ion api rework
Some of the work done for the ION 4.12+ ABI change
could be done in a cleaner fashion.

This tries to consolidate the logic so we have less
conditional switching depending on if we're using
legacy code or not, and also utilizes ion_alloc_fd()
to simplify the older ion_alloc usage.

Change-Id: I0891814663eaf26c7a9625331c2fc2db93700d0f
Signed-off-by: John Stultz <john.stultz@linaro.org>